requirements.txt
streamlit
python-dotenv
azure-search-documents==11.4.0
openai>=1.0.0
